+++ src/lib.rs
@@ -2,6 +2,7 @@ pub mod cache;
 pub mod cli;
 pub mod model;
 pub mod presentation;
+pub mod process;
 
 pub mod configure;
 pub mod dashboard;